chore: cleanup stale
