+++ apps/www/src/app/layout.tsx
@@ -12,7 +12,6 @@ import {
 } from "@ashgw/seo";
 import { fonts } from "@ashgw/ui";
 
-import { NavBar } from "~/app/components/misc/nav";
 import { env } from "@ashgw/env";
 import { TsrProvider } from "~/ts-rest/provider";
 
@@ -32,7 +31,6 @@ export default function RootLayout({ children }: PropsWithChildren) {
       <body className={fonts.atkinsonHyperlegible.className}>
         <JsonLd code={organizationJsonLd(siteUrl)} />
         <JsonLd code={websiteJsonLd(siteUrl)} />
-        <NavBar />
         <Providers site="www">
           <TsrProvider>{children}</TsrProvider>
         </Providers>
